src/third_party/WebKit/LayoutTests/external/wpt/fetch/sec-metadata/object.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Same-Origin object assert_equals: cause expected (undefined) undefined but got (string) "forced"
FAIL Same-Site object assert_equals: cause expected (undefined) undefined but got (string) "forced"
FAIL Cross-Site object assert_equals: cause expected (undefined) undefined but got (string) "forced"
Harness: the test ran to completion.
